Merge from v0.2.0-m6

+++ qupath-core/src/main/java/qupath/lib/measurements/NumericMeasurementList.java
@@ -157,7 +157,7 @@ class NumericMeasurementList {
 			// Read from map, if possible
 			if (map != null) {
 				Integer ind = map.get(name);
-				return ind == null ? -1 : ind;
+				return ind == null ? -1 : ind.intValue();
 			}
 			return names.indexOf(name);
 		}
